Result BmiService.java int to double

+++ src/BmiService.java
@@ -2,7 +2,7 @@ public class BmiService {
 
     public double calculate(double weight, double height) {
 
-        double result = weight / height / height;
+        int result = (int) (weight / height / height);
         return result;
     }
 }
